Refactor entire implementer tools configuration section -- just moving files around
