+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/config-subtree.component.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import EditableValue from "./editable-value.component";
-import { implementerToolsStore } from "../store";
+import { implementerToolsStore } from "../../store";
 import isEqual from "lodash-es/isEqual";
 import { Subtree } from "./layout/subtree.component";
