+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/config-tree.component.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./configuration.styles.scss";
+import styles from "./config-tree.styles.scss";
 import { Accordion, AccordionItem } from "carbon-components-react";
 import { ConfigTreeForModule } from "./config-tree-for-module.component";
 
@@ -17,7 +17,7 @@ export function ConfigTree({ config }: ConfigTreeProps) {
             const moduleConfig = config[moduleName];
             return Object.keys(moduleConfig).length ? (
               <AccordionItem
-                title={<h6 className={styles.moduleName}>{moduleName}</h6>}
+                title={<h6>{moduleName}</h6>}
                 className={styles.fullWidthAccordion}
                 key={`accordion-${moduleName}`}
               >
