+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/config-tree.styles.scss
@@ -1,0 +1,3 @@
+.fullWidthAccordion :global(.bx--accordion__content) {
+  padding-right: 0;
+}
