+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/description.component.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./description.styles.scss";
 import { useStore } from "@openmrs/esm-framework";
-import { implementerToolsStore } from "../store";
+import { implementerToolsStore } from "../../store";
 import { useTranslation } from "react-i18next";
 
 export function Description() {
