+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/description.styles.scss
@@ -1,4 +1,4 @@
-@import "../implementer-tools.styles.scss";
+@import "../../implementer-tools.styles.scss";
 
 .path {
   @include carbon--type-style("productive-heading-02");
