+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/display-value.scss
@@ -1,0 +1,5 @@
+@import "../../implementer-tools.styles.scss";
+
+.smallListCell {
+  padding: $spacing-03;
+}
