+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/display-value.tsx
@@ -1,6 +1,6 @@
 import { StructuredListCell, StructuredListRow } from "carbon-components-react";
 import React from "react";
-import styles from "./configuration.styles.scss";
+import styles from "./display-value.scss";
 
 export interface DisplayValueProps {
   value: any;
