+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/editable-value.styles.scss
@@ -1,4 +1,4 @@
-@import "../implementer-tools.styles.scss";
+@import "../../implementer-tools.styles.scss";
 
 .secretButton {
   background: none;
