+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/value-editor.scss
@@ -1,0 +1,5 @@
+@import "../../implementer-tools.styles.scss";
+
+.valueEditorButtons {
+  margin-top: $spacing-03;
+}
