+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/value-editors/concept-search.resource.tsx
