+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/value-editors/concept-search.styles.scss
