+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/value-editors/person-attribute-search.resource.tsx
