+++ packages/apps/esm-implementer-tools-app/src/configuration/interactive-editor/value-editors/uuid-search.scss
@@ -1,4 +1,4 @@
-@import "../../implementer-tools.styles.scss";
+@import "../../../implementer-tools.styles.scss";
 
 .activeUuid {
   @include carbon--type-style("body-short-01");
