+++ packages/apps/esm-implementer-tools-app/translations/en.json
@@ -1,14 +1,15 @@
 {
   "activeItemSourceText": "The current value comes from ",
   "backendModules": "Backend Modules",
-  "clearTemporaryConfig": "Clear Temporary Config",
+  "clearConfig": "Clear Local Config",
   "configuration": "Configuration",
   "devConfig": "Dev Config",
-  "downloadTemporaryConfig": "Download Temporary Config",
+  "downloadConfig": "Download Config",
   "edit": "Edit",
   "editValueButtonText": "Edit",
   "installedVersion": "Installed Version",
   "itemDescriptionSourceDefaultText": "The current value is the default.",
+  "jsonEditor": "JSON Editor",
   "missing": "Missing",
   "moduleName": "Module Name",
   "requiredVersion": "Required Version",
